+++ src/me/kstep/ucalc/InvertOp.java
@@ -4,6 +4,6 @@ class InvertOp extends UOperation {
     public int arity() { return 1; }
     public CharSequence name() { return "1/x"; }
     public void apply(UStack stack) {
-        stack.push(1 / stack.pop().floatValue());
+        stack.push(1 / stack.pop().doubleValue());
     }
 }
