+++ src/me/kstep/ucalc/LogOp.java
@@ -4,6 +4,6 @@ class LogOp extends UOperation {
     public int arity() { return 1; }
     public CharSequence name() { return "log"; }
     public void apply(UStack stack) {
-        stack.push(Math.log10(stack.pop().floatValue()));
+        stack.push(Math.log10(stack.pop().doubleValue()));
     }
 }
