+++ src/me/kstep/ucalc/SinOp.java
@@ -4,6 +4,6 @@ class SinOp extends UOperation {
     public int arity() { return 1; }
     public CharSequence name() { return "sin"; }
     public void apply(UStack stack) {
-        stack.push(Math.sin(stack.pop().floatValue()));
+        stack.push(Math.sin(stack.pop().doubleValue()));
     }
 }
